pip-requirements.txt
aiogram==3.2.0
asyncpg==0.29.0
python-dotenv==1.0.0
apscheduler==3.10.4
apscheduler_di==0.1.0
requests==2.31.0
